+++ data/taskList.txt
@@ -1,3 +1,4 @@
 T | 1 |  return book
 D | 0 |  homework | Feb 01 2023 2359
 E | 0 |  midterms | Mar 26 2023 1400 - Mar 26 2023 1600
+T | 1 |  play games
